+++ src/components/HelpScoutBeacon/BeaconHeadScripts.js
