+++ src/components/HelpScoutBeacon/HelpScoutBeacon.js
