+++ src/components/HelpScoutBeacon/IconQuestion.js
