+++ src/components/HelpScoutBeacon/logo.png
